Migrate all common types and functions to stdlib
Also introduces ordering of function overloads based on the
order in which they are specified as options or the order in
which they are merged together. This change ensures that the
type-checker's overload resolution checks will be deterministic.
