+++ cel/decls_test.go
@@ -21,9 +21,6 @@ import (
 	"strings"
 	"testing"
 
-	"google.golang.org/protobuf/proto"
-
-	"github.com/google/cel-go/checker"
 	chkdecls "github.com/google/cel-go/checker/decls"
 	"github.com/google/cel-go/common/decls"
 	"github.com/google/cel-go/common/functions"
@@ -521,332 +518,6 @@ func TestFunctionBinding(t *testing.T) {
 	}
 }
 
-func TestTypeToExprType(t *testing.T) {
-	tests := []struct {
-		in             *Type
-		out            *exprpb.Type
-		unidirectional bool
-	}{
-		{
-			in:  OpaqueType("vector", DoubleType, DoubleType),
-			out: chkdecls.NewAbstractType("vector", chkdecls.Double, chkdecls.Double),
-		},
-		{
-			in:  AnyType,
-			out: chkdecls.Any,
-		},
-		{
-			in:  BoolType,
-			out: chkdecls.Bool,
-		},
-		{
-			in:  BytesType,
-			out: chkdecls.Bytes,
-		},
-		{
-			in:  DoubleType,
-			out: chkdecls.Double,
-		},
-		{
-			in:  DurationType,
-			out: chkdecls.Duration,
-		},
-		{
-			in:  DynType,
-			out: chkdecls.Dyn,
-		},
-		{
-			in:  IntType,
-			out: chkdecls.Int,
-		},
-		{
-			in:  ListType(TypeParamType("T")),
-			out: chkdecls.NewListType(chkdecls.NewTypeParamType("T")),
-		},
-		{
-			in:  MapType(TypeParamType("K"), TypeParamType("V")),
-			out: chkdecls.NewMapType(chkdecls.NewTypeParamType("K"), chkdecls.NewTypeParamType("V")),
-		},
-		{
-			in:  NullType,
-			out: chkdecls.Null,
-		},
-		{
-			in:  ObjectType("google.type.Expr"),
-			out: chkdecls.NewObjectType("google.type.Expr"),
-		},
-		{
-			in:  StringType,
-			out: chkdecls.String,
-		},
-		{
-			in:  TimestampType,
-			out: chkdecls.Timestamp,
-		},
-		{
-			in:  TypeType,
-			out: chkdecls.NewTypeType(chkdecls.Dyn),
-		},
-		{
-			in:  UintType,
-			out: chkdecls.Uint,
-		},
-		{
-			in:  NullableType(BoolType),
-			out: chkdecls.NewWrapperType(chkdecls.Bool),
-		},
-		{
-			in:  NullableType(BytesType),
-			out: chkdecls.NewWrapperType(chkdecls.Bytes),
-		},
-		{
-			in:  NullableType(DoubleType),
-			out: chkdecls.NewWrapperType(chkdecls.Double),
-		},
-		{
-			in:  NullableType(IntType),
-			out: chkdecls.NewWrapperType(chkdecls.Int),
-		},
-		{
-			in:  NullableType(StringType),
-			out: chkdecls.NewWrapperType(chkdecls.String),
-		},
-		{
-			in:  NullableType(UintType),
-			out: chkdecls.NewWrapperType(chkdecls.Uint),
-		},
-		{
-			in:             ObjectType("google.protobuf.Any"),
-			out:            chkdecls.Any,
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Duration"),
-			out:            chkdecls.Duration,
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Timestamp"),
-			out:            chkdecls.Timestamp,
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Value"),
-			out:            chkdecls.Dyn,
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.ListValue"),
-			out:            chkdecls.NewListType(chkdecls.Dyn),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Struct"),
-			out:            chkdecls.NewMapType(chkdecls.String, chkdecls.Dyn),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.BoolValue"),
-			out:            chkdecls.NewWrapperType(chkdecls.Bool),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.BytesValue"),
-			out:            chkdecls.NewWrapperType(chkdecls.Bytes),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.DoubleValue"),
-			out:            chkdecls.NewWrapperType(chkdecls.Double),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.FloatValue"),
-			out:            chkdecls.NewWrapperType(chkdecls.Double),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Int32Value"),
-			out:            chkdecls.NewWrapperType(chkdecls.Int),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.Int64Value"),
-			out:            chkdecls.NewWrapperType(chkdecls.Int),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.StringValue"),
-			out:            chkdecls.NewWrapperType(chkdecls.String),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.UInt32Value"),
-			out:            chkdecls.NewWrapperType(chkdecls.Uint),
-			unidirectional: true,
-		},
-		{
-			in:             ObjectType("google.protobuf.UInt64Value"),
-			out:            chkdecls.NewWrapperType(chkdecls.Uint),
-			unidirectional: true,
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.in.String(), func(t *testing.T) {
-			got, err := TypeToExprType(tc.in)
-			if err != nil {
-				t.Fatalf("TypeToExprType(%v) failed: %v", tc.in, err)
-			}
-			if !proto.Equal(got, tc.out) {
-				t.Errorf("TypeToExprType(%v) returned %v, wanted %v", tc.in, got, tc.out)
-			}
-			if tc.unidirectional {
-				return
-			}
-			roundTrip, err := ExprTypeToType(got)
-			if err != nil {
-				t.Fatalf("ExprTypeToType(%v) failed: %v", got, err)
-			}
-			if !tc.in.IsType(roundTrip) {
-				t.Errorf("ExprTypeToType(%v) returned %v, wanted %v", got, roundTrip, tc.in)
-			}
-		})
-	}
-}
-
-func TestExprTypeToType(t *testing.T) {
-	tests := []struct {
-		in  *exprpb.Type
-		out *Type
-	}{
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Any"),
-			out: AnyType,
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Duration"),
-			out: DurationType,
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Timestamp"),
-			out: TimestampType,
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Value"),
-			out: DynType,
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.ListValue"),
-			out: ListType(DynType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Struct"),
-			out: MapType(StringType, DynType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.BoolValue"),
-			out: NullableType(BoolType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.BytesValue"),
-			out: NullableType(BytesType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.DoubleValue"),
-			out: NullableType(DoubleType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.FloatValue"),
-			out: NullableType(DoubleType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Int32Value"),
-			out: NullableType(IntType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.Int64Value"),
-			out: NullableType(IntType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.StringValue"),
-			out: NullableType(StringType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.UInt32Value"),
-			out: NullableType(UintType),
-		},
-		{
-			in:  chkdecls.NewObjectType("google.protobuf.UInt64Value"),
-			out: NullableType(UintType),
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.in.String(), func(t *testing.T) {
-			got, err := ExprTypeToType(tc.in)
-			if err != nil {
-				t.Fatalf("ExprTypeToType(%v) failed: %v", tc.in, err)
-			}
-			if !got.IsType(tc.out) {
-				t.Errorf("ExprTypeToType(%v) returned %v, wanted %v", tc.in, got, tc.out)
-			}
-		})
-	}
-}
-
-func TestExprTypeToTypeInvalid(t *testing.T) {
-	tests := []struct {
-		in  *exprpb.Type
-		out string
-	}{
-		{
-			in:  &exprpb.Type{},
-			out: "unsupported type",
-		},
-		{
-			in:  &exprpb.Type{TypeKind: &exprpb.Type_Primitive{}},
-			out: "unsupported primitive type",
-		},
-		{
-			in:  &exprpb.Type{TypeKind: &exprpb.Type_WellKnown{}},
-			out: "unsupported well-known type",
-		},
-		{
-			in:  chkdecls.NewListType(&exprpb.Type{}),
-			out: "unsupported type",
-		},
-		{
-			in:  chkdecls.NewMapType(&exprpb.Type{}, chkdecls.Dyn),
-			out: "unsupported type",
-		},
-		{
-			in:  chkdecls.NewMapType(chkdecls.Dyn, &exprpb.Type{}),
-			out: "unsupported type",
-		},
-		{
-			in:  chkdecls.NewAbstractType("bad", &exprpb.Type{}),
-			out: "unsupported type",
-		},
-		{
-			in:  &exprpb.Type{TypeKind: &exprpb.Type_Wrapper{}},
-			out: "unsupported primitive type",
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.in.String(), func(t *testing.T) {
-			_, err := ExprTypeToType(tc.in)
-			if err == nil || !strings.Contains(err.Error(), tc.out) {
-				t.Fatalf("ExprTypeToType(%v) got %v, wanted error %v", tc.in, err, tc.out)
-			}
-		})
-	}
-}
-
 func TestExprDeclToDeclaration(t *testing.T) {
 	size, err := ExprDeclToDeclaration(
 		chkdecls.NewFunction("size", chkdecls.NewOverload("size_string", []*exprpb.Type{chkdecls.String}, chkdecls.Int)),
@@ -1005,64 +676,5 @@ func testCompile(t testing.TB, env *Env, expr string, want any) {
 		if err == nil || want.Error() != err.Error() {
 			t.Errorf("prg.Eval() got error '%v', wanted '%v'", err, want)
 		}
-	}
-}
-
-func TestStdLibDeclarations(t *testing.T) {
-	stdlibFuncs, err := stdlibFuncDecls()
-	if err != nil {
-		t.Fatalf("stdlibFuncDecls() failed: %v", err)
-	}
-	_, err = NewCustomEnv(
-		Declarations(checker.StandardFunctions()...),
-		Declarations(stdlibFuncs...),
-	)
-	if err != nil {
-		t.Fatalf("NewCustomEnv() failed when combining stdlib functions and checker mirrors: %v", err)
-	}
-	stdlibFuncMap := make(map[string]*exprpb.Decl, len(stdlibFuncs))
-	for _, fn := range stdlibFuncs {
-		stdlibFuncMap[fn.GetName()] = fn
-	}
-	for _, fn := range checker.StandardFunctions() {
-		stdFn, found := stdlibFuncMap[fn.GetName()]
-		if !found {
-			t.Errorf("function not found in stdlib: %v", fn)
-		}
-		if len(stdFn.GetFunction().GetOverloads()) != len(fn.GetFunction().GetOverloads()) {
-			t.Errorf("stdlib function and checker function have different overload sets. std: %v, checker: %v", stdFn, fn)
-		}
-		stdOverloadMap := make(map[string]*exprpb.Decl_FunctionDecl_Overload)
-		for _, o := range stdFn.GetFunction().GetOverloads() {
-			stdOverloadMap[o.GetOverloadId()] = o
-		}
-		for _, o := range fn.GetFunction().GetOverloads() {
-			stdOverload, found := stdOverloadMap[o.GetOverloadId()]
-			if !found {
-				t.Errorf("overload not found in stdlib: %v", o)
-			}
-			if len(stdOverload.GetTypeParams()) != len(o.GetTypeParams()) {
-				t.Errorf("type params differ. std: %v, checker: %v", stdOverload, o)
-			}
-			if stdOverload.GetIsInstanceFunction() != o.GetIsInstanceFunction() {
-				t.Errorf("instance function settings differ. std: %v, checker: %v", stdOverload, o)
-			}
-			if len(stdOverload.GetParams()) != len(o.GetParams()) {
-				t.Errorf("arg counts differ. std: %v, checker: %v", stdOverload, o)
-			}
-			for i, p := range stdOverload.GetParams() {
-				if !proto.Equal(p, o.GetParams()[i]) {
-					t.Errorf("arg types differ. std: %v, checker: %v", stdOverload.GetResultType(), o.GetResultType())
-				}
-			}
-			if !proto.Equal(stdOverload.GetResultType(), o.GetResultType()) {
-				t.Errorf("result types differ. std: %v, checker: %v", stdOverload.GetResultType(), o.GetResultType())
-			}
-			delete(stdOverloadMap, o.GetOverloadId())
-		}
-		delete(stdlibFuncMap, fn.GetName())
-	}
-	for _, fn := range stdlibFuncMap {
-		t.Errorf("function in stdlib but not checker standard functions: %v", fn)
 	}
 }
