+++ common/decls/BUILD.bazel
@@ -13,10 +13,12 @@ go_library(
     ],
     importpath = "github.com/google/cel-go/common/decls",
     deps = [
+        "//checker/decls:go_default_library",
         "//common/functions:go_default_library",
         "//common/types:go_default_library",
         "//common/types/ref:go_default_library",
         "//common/types/traits:go_default_library",
+        "@org_golang_google_genproto_googleapis_api//expr/v1alpha1:go_default_library",
     ],
 )
 
@@ -28,9 +30,12 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
+        "//checker/decls:go_default_library",
         "//common/overloads:go_default_library",
         "//common/types:go_default_library",
         "//common/types/ref:go_default_library",
         "//common/types/traits:go_default_library",
+        "@org_golang_google_genproto_googleapis_api//expr/v1alpha1:go_default_library",
+        "@org_golang_google_protobuf//proto:go_default_library",
     ],
 )
