+++ common/decls/types_test.go
@@ -15,11 +15,17 @@
 package decls
 
 import (
+	"strings"
 	"testing"
 	"time"
 
+	"google.golang.org/protobuf/proto"
+
+	chkdecls "github.com/google/cel-go/checker/decls"
 	"github.com/google/cel-go/common/overloads"
 	"github.com/google/cel-go/common/types"
+
+	exprpb "google.golang.org/genproto/googleapis/api/expr/v1alpha1"
 )
 
 func TestTypeString(t *testing.T) {
@@ -36,8 +42,20 @@ func TestTypeString(t *testing.T) {
 			out: "map(uint, double)",
 		},
 		{
-			in:  NullableType(BoolType),
+			in:  BoolType,
 			out: "bool",
+		},
+		{
+			in:  DynType,
+			out: "dyn",
+		},
+		{
+			in:  NullType,
+			out: "null_type",
+		},
+		{
+			in:  NullableType(BoolType),
+			out: "wrapper(bool)",
 		},
 		{
 			in:  OptionalType(ListType(StringType)),
@@ -49,11 +67,11 @@ func TestTypeString(t *testing.T) {
 		},
 		{
 			in:  ObjectType("google.protobuf.Int32Value"),
-			out: "int",
+			out: "wrapper(int)",
 		},
 		{
 			in:  ObjectType("google.protobuf.UInt32Value"),
-			out: "uint",
+			out: "wrapper(uint)",
 		},
 		{
 			in:  ObjectType("google.protobuf.Value"),
@@ -120,6 +138,35 @@ func TestTypeIsType(t *testing.T) {
 	for _, tst := range tests {
 		if tst.t1.IsType(tst.t2) != tst.isType {
 			t.Errorf("%v.IsType(%v) got %v, wanted %v", tst.t1, tst.t2, !tst.isType, tst.isType)
+		}
+	}
+}
+
+func TestTypeTypeVariable(t *testing.T) {
+	tests := []struct {
+		t *Type
+		v *VariableDecl
+	}{
+		{
+			t: AnyType,
+			v: NewVariable("google.protobuf.Any", TypeTypeWithParam(AnyType)),
+		},
+		{
+			t: DynType,
+			v: NewVariable("dyn", TypeTypeWithParam(DynType)),
+		},
+		{
+			t: ObjectType("google.protobuf.Int32Value"),
+			v: NewVariable("int", TypeTypeWithParam(NullableType(IntType))),
+		},
+		{
+			t: ObjectType("google.protobuf.Int32Value"),
+			v: NewVariable("int", TypeTypeWithParam(NullableType(IntType))),
+		},
+	}
+	for _, tst := range tests {
+		if !tst.t.TypeVariable().DeclarationEquals(tst.v) {
+			t.Errorf("got not equal %v.Equals(%v)", tst.t.TypeVariable(), tst.v)
 		}
 	}
 }
@@ -233,5 +280,415 @@ func TestTypeIsAssignableRuntimeType(t *testing.T) {
 	if MapType(StringType, DynType).IsAssignableRuntimeType(
 		types.DefaultTypeAdapter.NativeToValue(map[int64]time.Duration{1: time.Duration(1)})) {
 		t.Error("map(string, dyn) must not be assignable to map(int, duration) at runtime")
+	}
+}
+
+func TestTypeToExprType(t *testing.T) {
+	tests := []struct {
+		in             *Type
+		out            *exprpb.Type
+		unidirectional bool
+	}{
+		{
+			in:  OpaqueType("vector", DoubleType, DoubleType),
+			out: chkdecls.NewAbstractType("vector", chkdecls.Double, chkdecls.Double),
+		},
+		{
+			in:  AnyType,
+			out: chkdecls.Any,
+		},
+		{
+			in:  BoolType,
+			out: chkdecls.Bool,
+		},
+		{
+			in:  BytesType,
+			out: chkdecls.Bytes,
+		},
+		{
+			in:  DoubleType,
+			out: chkdecls.Double,
+		},
+		{
+			in:  DurationType,
+			out: chkdecls.Duration,
+		},
+		{
+			in:  DynType,
+			out: chkdecls.Dyn,
+		},
+		{
+			in:  IntType,
+			out: chkdecls.Int,
+		},
+		{
+			in:  ListType(TypeParamType("T")),
+			out: chkdecls.NewListType(chkdecls.NewTypeParamType("T")),
+		},
+		{
+			in:  MapType(TypeParamType("K"), TypeParamType("V")),
+			out: chkdecls.NewMapType(chkdecls.NewTypeParamType("K"), chkdecls.NewTypeParamType("V")),
+		},
+		{
+			in:  NullType,
+			out: chkdecls.Null,
+		},
+		{
+			in:  ObjectType("google.type.Expr"),
+			out: chkdecls.NewObjectType("google.type.Expr"),
+		},
+		{
+			in:  StringType,
+			out: chkdecls.String,
+		},
+		{
+			in:  TimestampType,
+			out: chkdecls.Timestamp,
+		},
+		{
+			in:  TypeType,
+			out: chkdecls.NewTypeType(nil),
+		},
+		{
+			in:  UintType,
+			out: chkdecls.Uint,
+		},
+		{
+			in:  NullableType(BoolType),
+			out: chkdecls.NewWrapperType(chkdecls.Bool),
+		},
+		{
+			in:  NullableType(BytesType),
+			out: chkdecls.NewWrapperType(chkdecls.Bytes),
+		},
+		{
+			in:  NullableType(DoubleType),
+			out: chkdecls.NewWrapperType(chkdecls.Double),
+		},
+		{
+			in:  NullableType(IntType),
+			out: chkdecls.NewWrapperType(chkdecls.Int),
+		},
+		{
+			in:  NullableType(StringType),
+			out: chkdecls.NewWrapperType(chkdecls.String),
+		},
+		{
+			in:  NullableType(UintType),
+			out: chkdecls.NewWrapperType(chkdecls.Uint),
+		},
+		{
+			in:  TypeTypeWithParam(TypeTypeWithParam(DynType)),
+			out: chkdecls.NewTypeType(chkdecls.NewTypeType(chkdecls.Dyn)),
+		},
+		{
+			in:             ObjectType("google.protobuf.Any"),
+			out:            chkdecls.Any,
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Duration"),
+			out:            chkdecls.Duration,
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Timestamp"),
+			out:            chkdecls.Timestamp,
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Value"),
+			out:            chkdecls.Dyn,
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.ListValue"),
+			out:            chkdecls.NewListType(chkdecls.Dyn),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Struct"),
+			out:            chkdecls.NewMapType(chkdecls.String, chkdecls.Dyn),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.BoolValue"),
+			out:            chkdecls.NewWrapperType(chkdecls.Bool),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.BytesValue"),
+			out:            chkdecls.NewWrapperType(chkdecls.Bytes),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.DoubleValue"),
+			out:            chkdecls.NewWrapperType(chkdecls.Double),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.FloatValue"),
+			out:            chkdecls.NewWrapperType(chkdecls.Double),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Int32Value"),
+			out:            chkdecls.NewWrapperType(chkdecls.Int),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.Int64Value"),
+			out:            chkdecls.NewWrapperType(chkdecls.Int),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.StringValue"),
+			out:            chkdecls.NewWrapperType(chkdecls.String),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.UInt32Value"),
+			out:            chkdecls.NewWrapperType(chkdecls.Uint),
+			unidirectional: true,
+		},
+		{
+			in:             ObjectType("google.protobuf.UInt64Value"),
+			out:            chkdecls.NewWrapperType(chkdecls.Uint),
+			unidirectional: true,
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.in.String(), func(t *testing.T) {
+			got, err := TypeToExprType(tc.in)
+			if err != nil {
+				t.Fatalf("TypeToExprType(%v) failed: %v", tc.in, err)
+			}
+			if !proto.Equal(got, tc.out) {
+				t.Errorf("TypeToExprType(%v) returned %v, wanted %v", tc.in, got, tc.out)
+			}
+			if tc.unidirectional {
+				return
+			}
+			roundTrip, err := ExprTypeToType(got)
+			if err != nil {
+				t.Fatalf("ExprTypeToType(%v) failed: %v", got, err)
+			}
+			if !tc.in.IsType(roundTrip) {
+				t.Errorf("ExprTypeToType(%v) returned %v, wanted %v", got, roundTrip, tc.in)
+			}
+		})
+	}
+}
+
+func TestTypeToExprTypeInvalid(t *testing.T) {
+	tests := []struct {
+		in  *Type
+		out string
+	}{
+		{
+			in:  &Type{Kind: ListKind, runtimeType: types.ListType},
+			out: "invalid list",
+		},
+		{
+			in: &Type{
+				Kind: ListKind,
+				Parameters: []*Type{
+					{Kind: MapKind, runtimeType: types.MapType},
+				},
+				runtimeType: types.ListType,
+			},
+			out: "invalid map",
+		},
+		{
+			in:  &Type{Kind: MapKind, runtimeType: types.MapType},
+			out: "invalid map",
+		},
+		{
+			in: &Type{
+				Kind: MapKind,
+				Parameters: []*Type{
+					StringType,
+					{Kind: MapKind, runtimeType: types.MapType},
+				},
+				runtimeType: types.MapType,
+			},
+			out: "invalid map",
+		},
+		{
+			in: &Type{
+				Kind: MapKind,
+				Parameters: []*Type{
+					{Kind: MapKind, runtimeType: types.MapType},
+					StringType,
+				},
+				runtimeType: types.MapType,
+			},
+			out: "invalid map",
+		},
+		{
+			in: &Type{
+				Kind:        TypeKind,
+				Parameters:  []*Type{{Kind: ListKind, runtimeType: types.ListType}},
+				runtimeType: types.TypeType,
+			},
+			out: "invalid list",
+		},
+		{
+			in: OpaqueType("bad_list", &Type{
+				Kind:        ListKind,
+				runtimeType: types.ListType,
+			}),
+			out: "invalid list",
+		},
+		{
+			in:  &Type{},
+			out: "missing type conversion",
+		},
+	}
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.in.String(), func(t *testing.T) {
+			_, err := TypeToExprType(tc.in)
+			if err == nil || !strings.Contains(err.Error(), tc.out) {
+				t.Fatalf("TypeToExprType(%v) got %v, wanted error %v", tc.in, err, tc.out)
+			}
+		})
+	}
+}
+
+func TestExprTypeToType(t *testing.T) {
+	tests := []struct {
+		in  *exprpb.Type
+		out *Type
+	}{
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Any"),
+			out: AnyType,
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Duration"),
+			out: DurationType,
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Timestamp"),
+			out: TimestampType,
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Value"),
+			out: DynType,
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.ListValue"),
+			out: ListType(DynType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Struct"),
+			out: MapType(StringType, DynType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.BoolValue"),
+			out: NullableType(BoolType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.BytesValue"),
+			out: NullableType(BytesType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.DoubleValue"),
+			out: NullableType(DoubleType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.FloatValue"),
+			out: NullableType(DoubleType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Int32Value"),
+			out: NullableType(IntType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.Int64Value"),
+			out: NullableType(IntType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.StringValue"),
+			out: NullableType(StringType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.UInt32Value"),
+			out: NullableType(UintType),
+		},
+		{
+			in:  chkdecls.NewObjectType("google.protobuf.UInt64Value"),
+			out: NullableType(UintType),
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.in.String(), func(t *testing.T) {
+			got, err := ExprTypeToType(tc.in)
+			if err != nil {
+				t.Fatalf("ExprTypeToType(%v) failed: %v", tc.in, err)
+			}
+			if !got.IsType(tc.out) {
+				t.Errorf("ExprTypeToType(%v) returned %v, wanted %v", tc.in, got, tc.out)
+			}
+		})
+	}
+}
+
+func TestExprTypeToTypeInvalid(t *testing.T) {
+	tests := []struct {
+		in  *exprpb.Type
+		out string
+	}{
+		{
+			in:  &exprpb.Type{},
+			out: "unsupported type",
+		},
+		{
+			in:  &exprpb.Type{TypeKind: &exprpb.Type_Primitive{}},
+			out: "unsupported primitive type",
+		},
+		{
+			in:  &exprpb.Type{TypeKind: &exprpb.Type_WellKnown{}},
+			out: "unsupported well-known type",
+		},
+		{
+			in:  chkdecls.NewListType(&exprpb.Type{}),
+			out: "unsupported type",
+		},
+		{
+			in:  chkdecls.NewMapType(&exprpb.Type{}, chkdecls.Dyn),
+			out: "unsupported type",
+		},
+		{
+			in:  chkdecls.NewMapType(chkdecls.Dyn, &exprpb.Type{}),
+			out: "unsupported type",
+		},
+		{
+			in:  chkdecls.NewAbstractType("bad", &exprpb.Type{}),
+			out: "unsupported type",
+		},
+		{
+			in:  &exprpb.Type{TypeKind: &exprpb.Type_Wrapper{}},
+			out: "unsupported primitive type",
+		},
+		{
+			in:  &exprpb.Type{TypeKind: &exprpb.Type_Type{Type: &exprpb.Type{TypeKind: &exprpb.Type_Function{}}}},
+			out: "unsupported type",
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.in.String(), func(t *testing.T) {
+			_, err := ExprTypeToType(tc.in)
+			if err == nil || !strings.Contains(err.Error(), tc.out) {
+				t.Fatalf("ExprTypeToType(%v) got %v, wanted error %v", tc.in, err, tc.out)
+			}
+		})
 	}
 }
